healthcheckreport/requirements.txt
azure_communication_chat==1.1.0
azure_communication_email==1.0.0
azure_communication_sms==1.0.1
azure_mgmt_compute==30.0.0
azure_mgmt_core==1.4.0
email_client==1.5.2
typing_extensions==4.7.1
azure-core==1.28.0
azure-identity==1.13.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
fabric2==2.7.1
idna==3.4
invoke==1.7.3
msal==1.21.0
msal-extensions==1.0.0
paramiko==3.1.0
pathlib2==2.3.7.post1
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pywin32==305
requests==2.28.2
six==1.16.0
urllib3==1.26.15
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.94.0
google-cloud-compute==1.11.0